move eval to tests folder
