+++ tests/eval/indexManagementTextSearchIndexCreation.eval.ts
